requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
pydantic
ultralytics
torch
torchvision
torchaudio
python-dotenv